+++ package.json
@@ -32,9 +32,6 @@
     "most": "^1.7.2",
     "ws": "^3.3.3"
   },
-  "optionalDependencies": {
-    "chalk": "^2.3.0"
-  },
   "devDependencies": {
     "@types/async": "2.0.47",
     "@types/chai": "^4.1.0",
