+++ packages/util-types/src/index.ts
@@ -1,1 +1,3 @@
-export * from "./types";
+export * from "./event_emitter";
+export * from "./helpers";
+export * from "./json";
