+++ packages/util-types/src/json.ts
@@ -14,16 +14,3 @@ export const isJsonObject = (jsonType: JsonType): jsonType is JsonObject => {
     && typeof jsonType === "object"
     && !isJsonArray(jsonType); // arrays are also of type object
 };
-
-// ---------------
-
-export type Nullable<T> = T | undefined;
-export type Purify<T extends string> = { [P in T]: T; }[T];
-/**
- * Make all properties of `T` optional.
- *
- * @see https://github.com/Microsoft/TypeScript/issues/15012#issuecomment-346499713
- */
-export type Optional<T> = {
-  [P in Purify<keyof T>]: Nullable<T[P]>;
-};
